+++ config/webpack/webpack.config.js
@@ -112,7 +112,7 @@ const webpackConfig = (options, env, argv) => {
         children: false,
       },
     },
-    devtool: 'cheap-source-map',
+    devtool: 'eval-source-map',
     resolveLoader: {
       modules: [path.resolve(path.join(rootPath, 'node_modules'))],
     },